pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.22.0",
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "python-decouple>=3.8",
    "sqlalchemy>=2.0.43",
]
